ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
google-api-python-client==2.108.0
supabase==2.0.2
transformers==4.36.0
torch==2.1.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
nltk==3.8.1
textblob==0.17.1
vaderSentiment==3.3.2
celery==5.3.4
redis==5.0.1 